Refactor: merge auth store into user store and relocate brand upload to services
