+++ src/index.css
@@ -25,6 +25,7 @@
 	line-height: 1.5;
 	font-weight: 400;
 
+	
 	color-scheme: light dark;
 	background-color: rgba(255, 255, 255, 0.87);
 	color: #242424;
@@ -34,7 +35,8 @@
 	-webkit-font-smoothing: antialiased;
 	-moz-osx-font-smoothing: grayscale;
 
-	--radius: 0.625rem;
+	--radius-md: 11px
+	--radius-lg: 30px;
 
 	--background: oklch(1 0 0);
 	--foreground: oklch(0.145 0 0);
@@ -45,15 +47,15 @@
 	--primary: #29b6f6;
 	--primary-hover: #039be5;
 	--primary-foreground: oklch(0.985 0 0);
-	--secondary: #00acc1;
-	--secondary-hover: #f0f0f0;
-	--secondary-foreground: oklch(0.205 0 0);
-	--muted: oklch(0.97 0 0);
+	--secondary: #ED775A; /* رنگ جدید */
+  	--secondary-hover: #e36a4e; /* یک سایه تیره‌تر برای hover */
+  	--secondary-foreground: #ffffff; /* متن سفید روی این رنگ */
+	--muted: #f0f0f0;
 	--muted-foreground: oklch(0.556 0 0);
 	--accent: oklch(0.97 0 0);
 	--accent-foreground: oklch(0.205 0 0);
 	--destructive: oklch(0.577 0.245 27.325);
-	--border: oklch(0.922 0 0);
+	--border: #f0f0f0;
 	--input: oklch(0.922 0 0);
 	--ring: oklch(0.708 0 0);
 	--chart-1: oklch(0.646 0.222 41.116);
@@ -75,8 +77,8 @@
 /* Theme setup */
 @theme inline {
 	--radius-sm: calc(var(--radius) - 4px);
-	--radius-md: calc(var(--radius) - 2px);
-	--radius-lg: var(--radius);
+	--radius-md: 11px
+	--radius-lg: 30px;
 	--radius-xl: calc(var(--radius) + 4px);
 
 	--color-background: var(--background);
