--- src/store/useAuthStore.ts
@@ -1,22 +1,0 @@
-import { create } from "zustand";
-import { persist, createJSONStorage } from "zustand/middleware";
-
-interface AuthState {
-  token: string | null;
-  setToken: (token: string | null) => void;
-}
-
-const useAuthStore = create<AuthState>()(
-  persist(
-    (set) => ({
-      token: null,
-      setToken: (token) => set({ token }),
-    }),
-    {
-      name: "auth-storage",
-      storage: createJSONStorage(() => sessionStorage), // یا localStorage
-    }
-  )
-);
-
-export default useAuthStore;
