+++ src/store/userStore/userStore.ts
@@ -1,22 +1,25 @@
-// src/stores/useProfileStore.ts
-
 import { create } from "zustand";
 import { persist, createJSONStorage } from "zustand/middleware";
 import type { UserState } from "../../types/userTypes";
 
 const useUserStore = create<UserState>()(
-	persist(
-		(set) => ({
-			username: null,
+  persist(
+    (set) => ({
+      // state
+      token: null,
+      username: null,
 
-			setUsername: (username: string) =>
-				set((prev) => ({ ...prev, username })),
-		}),
-		{
-			name: "profile-storage",
-			storage: createJSONStorage(() => sessionStorage),
-		}
-	)
+      // actions
+      setToken: (token) => set({ token }),
+
+      setUsername: (username: string) =>
+        set((prev) => ({ ...prev, username })),
+    }),
+    {
+      name: "profile-storage",
+      storage: createJSONStorage(() => sessionStorage),
+    }
+  )
 );
 
 export default useUserStore;
