+++ src/types/userTypes.ts
@@ -1,4 +1,7 @@
 export interface UserState {
-	username: string | null;
-	setUsername: (username: string) => void;
+  token: string | null;
+  username: string | null;
+
+  setToken: (token: string | null) => void;
+  setUsername: (username: string) => void;
 }
